+++ source/commands/petition.js
@@ -1,6 +1,6 @@
 const Command = require('../classes/Command.js');
-const { getTopicNames, findTopicId } = require('../engines/channelEngine.js');
-const { joinChannel, checkPetition } = require('../helpers.js');
+const { getTopicNames } = require('../engines/channelEngine.js');
+const { checkPetition } = require('../helpers.js');
 
 const options = [
 	{ type: "String", name: "topic-name", description: "Make sure the topic doesn't already exist", required: true, choices: [] }
@@ -12,15 +12,12 @@ module.exports = new Command("petition", "Petition for a topic text channel", "n
  * @param {import('discord.js').Interaction} interaction */
 module.exports.execute = (interaction) => {
 	const topicName = interaction.options.getString("topic-name").toLowerCase();
-	if (getTopicNames().includes(topicName)) {
-		interaction.guild.channels.fetch(findTopicId(topicName)).then(textChannel => {
-			joinChannel(textChannel, interaction.user);
-		})
-		interaction.reply({ content: `A channel for ${topicName} already exists, you've been added to that channel.`, ephemeral: true })
-			.catch(console.error);
-	} else {
+	if (!getTopicNames().includes(topicName)) {
 		checkPetition(interaction.guild, topicName, interaction.user);
 		interaction.reply(`Your petition for **${topicName}** has been recorded!`)
+			.catch(console.error);
+	} else {
+		interaction.reply({ content: `A channel for ${topicName} already exists, you can use the Channels & Roles Browser to make it visible.`, ephemeral: true })
 			.catch(console.error);
 	}
 }